+++ client/app/utils/api-path-helpers.js
@@ -11,6 +11,10 @@ export function uploadManuscriptPath(taskId) {
   return `/api/tasks/${taskId}/upload_manuscript`;
 }
 
+export function uploadSourceFilePath(taskId) {
+  return `/api/tasks/${taskId}/upload_sourcefile`;
+}
+
 export function filteredUsersPath(paperId) {
   return `/api/filtered_users/users/${paperId}`;
 }
@@ -24,9 +28,16 @@ export function newDiscussionUsersPath(paperId) {
 }
 
 export function paperDownloadPath({paperId, versionedTextId, format}) {
-  let path = `/api/paper_downloads/${paperId}?export_format=${format}`;
+  let path = `/api/paper_downloads/${paperId}`;
+  let params = {};
+  if (format) {
+    params.export_format = format;
+  }
   if (versionedTextId) {
-    path += `&versioned_text_id=${versionedTextId}`;
+    params.versioned_text_id = versionedTextId;
+  }
+  if (!_.isEmpty(params)) {
+    path += `?${$.param(params)}`;
   }
   return path;
 }