+++ engines/tahi_standard_tasks/client/app/components/upload-source-file.js
@@ -1,26 +1,25 @@
-import TaskComponent from 'tahi/pods/components/task-base/component';
 import FileUploadMixin from 'tahi/mixins/file-upload';
+import { uploadSourceFilePath } from 'tahi/utils/api-path-helpers';
 import fontAwesomeFiletypeClass from 'tahi/lib/font-awesome-fyletype-class';
-import { uploadManuscriptPath } from 'tahi/utils/api-path-helpers';
 import Ember from 'ember';
 
-export default TaskComponent.extend(FileUploadMixin, {
+export default Ember.Component.extend(FileUploadMixin, {
   progress: 0,
   showProgress: true,
 
-  pdfAllowed: Ember.computed.reads('task.paper.journal.pdfAllowed'),
+  pdfAllowed: false,
 
   progressBarStyle: Ember.computed('progress', function() {
     return Ember.String.htmlSafe('width:' + this.get('progress') + '%');
   }),
 
-  manuscriptUploadUrl: Ember.computed('task.id', function() {
-    return uploadManuscriptPath(this.get('task.id'));
+  sourceFileUploadUrl: Ember.computed('task.id', function() {
+    return uploadSourceFilePath(this.get('task.id'));
   }),
 
   fileTypeClass: Ember.computed('filename', 'task.paper.file.filename', function(){
     let uploaded = this.get('manuscriptfileUploaded');
-    return fontAwesomeFiletypeClass(this.get(uploaded ? 'filename' : 'task.paper.file.filename'));
+    return fontAwesomeFiletypeClass(this.get(uploaded ? 'filename' : 'task.paper.sourcefile.filename'));
   }),
 
   actions: {
@@ -51,10 +50,8 @@ export default TaskComponent.extend(FileUploadMixin, {
 
     uploadFinished(data, filename, s3Url) {
       this.uploadFinished(data, filename);
-      this.get('store').pushPayload(data);
-
       this.get('task').save();
-      this.set('manuscriptUploaded', true);
+      this.set('sourcefileUploaded', true);
       this.set('s3Url', s3Url);
       this.set('filename', filename);
     }